Move all metadata out
