+++ src/query/service/src/sql/planner/mod.rs
@@ -37,11 +37,9 @@ pub use binder::Binder;
 pub use binder::ColumnBinding;
 pub use binder::Visibility;
 use common_catalog::catalog::CatalogManager;
+use common_planner::Metadata;
+use common_planner::MetadataRef;
 pub use metadata::find_smallest_column;
-pub use metadata::ColumnEntry;
-pub use metadata::Metadata;
-pub use metadata::MetadataRef;
-pub use metadata::TableEntry;
 pub use semantic::normalize_identifier;
 pub use semantic::IdentifierNormalizer;
 pub use semantic::NameResolutionContext;
@@ -98,7 +96,7 @@ impl Planner {
                 let (stmt, format) = parse_sql(&tokens, sql_dialect, &backtrace)?;
 
                 // Step 3: Bind AST with catalog, and generate a pure logical SExpr
-                let metadata = Arc::new(RwLock::new(Metadata::create()));
+                let metadata = Arc::new(RwLock::new(Metadata::default()));
                 let name_resolution_ctx = NameResolutionContext::try_from(settings.as_ref())?;
                 let binder = Binder::new(
                     self.ctx.clone(),
