+++ src/query/service/src/sql/planner/optimizer/heuristic/subquery_rewriter.rs
@@ -23,6 +23,7 @@ use common_exception::ErrorCode;
 use common_exception::Result;
 use common_functions::aggregates::AggregateFunctionFactory;
 use common_planner::IndexType;
+use common_planner::MetadataRef;
 
 use crate::sql::binder::ColumnBinding;
 use crate::sql::binder::Visibility;
@@ -50,7 +51,6 @@ use crate::sql::plans::Scalar;
 use crate::sql::plans::ScalarItem;
 use crate::sql::plans::SubqueryExpr;
 use crate::sql::plans::SubqueryType;
-use crate::sql::MetadataRef;
 use crate::sql::ScalarExpr;
 
 pub enum UnnestResult {
