+++ src/query/service/src/sql/planner/optimizer/util.rs
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_planner::MetadataRef;
+
 use super::SExpr;
 use crate::sql::plans::JoinType;
 use crate::sql::plans::RelOperator;
-use crate::sql::MetadataRef;
 
 /// Check if a query will read data from local tables(e.g. system tables).
 pub fn contains_local_table_scan(s_expr: &SExpr, metadata: &MetadataRef) -> bool {
